scraper_q: fix error on no-image links
In certain error conditions (namely, if we fail to fetch either an image or
html from the link, or there are no images in the html),
`_find_thumbnail_image()` returns `None`, which has been resulting in an
exception afterwards.  The other place we do this doesn't need the same check,
because we shouldn't get `None` as the value we're passing into it (and if we
do, that's an *unexpected* error).