Multi API: Allow for auto-sluggification of name
This updates some multi endpoints to allow for passing in or using
"display_name" instead of a multi path when creating a new multi. In such
a case, the display_name will be converted into a "slug" for the multi (a
string valid for use as a multi name), and that slug will be used.

This affects:

* POST /api/multi/
* POST /api/multi/rename
* POST /api/multi/copy

(PUT /api/multi/{path} is not affected, as it requires the caller to have
a slug in mind already for an existing or new multi)