Add new Q&A-oriented sort type
None of our current sort types work terribly well for Q&A threads.  So, we're
adding a new one that takes into account not only the votes on top-level
comments, but the votes on any OP responses, too, as well as the length of both
comments.

We're currently only interested in using this through API consumers, so while
it's available on reddit.com, the controls to turn it on are not exposed.