Reduce gzip threshold for main caches to 1400 bytes.
The goal here is to reduce network traffic to the main cache nodes.
This will also affect the size of the objects in the cache which could
make things better by fitting more things in cache, or make things worse
by oversubscribing specific slabs. We should watch network traffic on
the cache nodes, hit rates, and eviction rates in the period after
rolling this out. Additionally, this will increase the amount of
gzipping app servers deal with, so response times and CPU should be
monitored as well.

I chose 1400 bytes to be a fudge-factor smaller than the MTU; since the
goal here is network efficiency and not cache memory space, going
smaller than a single packet isn't terribly important.