analytics.js: Ensure GA callbacks are called.
If for whatever reason ga fails to call the callback in a
reasonable amount of time (500ms) then fire it off anyway. One
instance where this was biting us was, an addon for firefox
(ghostery) was stubbing out `_gaq.push` with a noop, causing any
functions pushed into it to never be fired.