Account.needs_captcha: add hook, use live_config
This allows the minimum amount of karma needed to be exempt from the
captcha to be modified via live config. In addition, it adds the
ability to set a comment karma minimum, where previously it was required
to get link karma in order to be exempt from the captcha.

A hook has also been added to the function for private-code purposes.