Remove unnecessary global margin-top from checkbox/radio buttons.
It appears this was [added back in
2008](https://github.com/reddit/reddit-public/commit/58fd1e858210fa0b2f1a36d103e7b6cd5a364c41)
as part of other work, but this is a pretty odd piece of style
to have in the global space, and additionally it's overridden
later for any input within a form in
https://github.com/reddit/reddit-public/blob/master/r2/r2/public/static/css/reddit.less#L4166
anyway.

I briefly dug through the code looking for any places that
we've got checkbox inputs outside of forms. The only places I
found so far were multis (overrode the style anyway) and the
embed modal (which I'm fixing).