First cut at hosting multiple-size thumbnails
We'd like to display multiple sizes of thumbnails for the mobile apps, rather
than just the one size we currently use.  The first step to this is storing the
source image, rather than just the one resized version, so we can create all
the differently-sized versions from it.

The resizing is implemented using the provider system because that gives us
flexibility - just as open-source and dev installs can use local file storage
rather than S3, so they can use a no-op resizer or one that pulls in correct
image sizes from a placeholder image site.

To note: none of this code should actually be running right now; it still needs
to be hooked up to the rest of the system.