vote: Use LastModified to cull some prequeued_votes lookups.
prequeued_votes is meant to hide the fact that votes are processed
asynchronously when a user refreshes a page right after voting (i.e. the
vote arrow will remain colored even if it wasn't processed fully yet on
the backend.)

Now that LastModified for votes is written in-request rather than in the
queue, we can take advantage of this to reduce the number of items we
have to look up prequeued_votes for.

A new configurable, vote_queue_grace_period, specifies what amount of
queue lag we're willing to paper over.  If the user hasn't voted in that
grace period, we just assume that the vote has been processed, skip
prequeued_votes lookups for it and go straight to Cassandra.
Additionally, if we are going to do lookups, we'll skip lookups for
items that were created since the last time the user voted like we were
already doing at the DenormalizedRelation layer.

This will likely shift some reads to Cassandra so we should keep an eye
on that.