CommentBuilder: rehide deleted no-children comments
As part of bbc4ac36, I split an iteration through comments in `CommentBuilder`
into two loops.  Unfortunately, we have logic at the top of the first loop that
skips onto the next item if the one we're looking at is deleted and has no
children, and that logic was being bypassed in the newly-created second loop.
This littered deleted comments everywhere, which is particularly problematic
for heavily-moderated subreddits like /r/askscience.

Now we don't do that. :)