Q&A sort: feature flag
We didn't previously gate this behind a feature flag because there is code
built to hide sorts from the various menus, so we could just tell people to
flag it on with a url parameter.

However, we now would like the Q&A sort to appear in the primary sort menu, and
the recommended sort menu, for certain groups of people so as to make testing
easier.