Q&A sort: update parents as necessary
While thinking about how annoying it'd be to do various things with the sort,
since it's built to process each comment individually, I realized there was a
bug with the current implementation.

Like with most things on reddit, we try really hard to do all of our sorting
as-needed, and cache the results as long as possible.  When something happens
to a comment that would change its sort value (eg it's posted or voted on), we
throw it into a queue of things to update.

The problem here is that Q&A sort has wider requirements for calculating a
sort; namely, we need to look at a comment's children.  Here's an example:

User A posts a question in an AMA.  Since there are no answers from the OP, the
score of this question is just taking into the accounts the votes and length of
this one comment.  When OP responds, the sort value on User A's comment should
also now take into account that it has an OP reply.  However, because we only
updated the sort value for the changed comment (OP's), that is not the case.

Practically speaking, this has a fairly small effect; whenever someone votes on
User A's comment, its value is recalculated, taking into account its children.
But until this happens, which can be a while in small threads, the sort would
look weird (most noticeably by having an OP-responded question not in the top
part of the thread).

There are a number of entry points into the functions calculating sort values.
I picked the "highest" one that's in common - that is, so we only have to have
this logic in *one* place, rather than scattered across multiple call
locations, since that would inevitably lead to somewhere being missed.  Given
the nature of sort values and caching, these bugs are incredibly hard to track
down.