Image previews: work around s3 bucket sharding
To help pipeline requests, we host images in multiple s3 buckets, each pointed
at by a different subdomain.  The bucket names are the domains that point to
them, which we take advantage of by setting `s3_media_direct` to false in
production to use the first part of the s3 path (the bucket) as the domain name
for the request.

With the introduction of imgix, however, we have another endpoint that proxies
through to our images.  We have had it pointing to one of those domains, which
worked for some images... but not the ones in a different bucket.

If we keep the bucket name in the path, we can pass that on to
`s3.amazonaws.com`, and things will work out just fine.  However, this is a
temporary measure, given that it opens us up to others using our imgix domain
for their own images.  More detail on this is given in the code comments.