Scraper: gracefully handle no embedly thumbnail
While poking around in the `scraper_q` logs, I saw another unhandled exception
I had erroneously lumped in with the errors in #3362.

It turns out that the `thumbnail_url` attribute in embedly's oembed responses
is actually [optional][0], and so sometimes we've been trying to fetch a url of
`None`.

Now we simply check for a missing url and abort early if that's the case.

[0]: http://embed.ly/docs/api/embed/endpoints/1/oembed#response