Image previews: fix old urls
In #3376 we changed the url structure of image previews, and we're changing it
again in #3419.  This has lead to some amount of trouble for the mobile folks,
since they're putting in place images that are 404ing.

This script should find all preview images that have been uploaded thus far,
move them to the appropriate place, and save an updated and correct storage url
in every Link that uses them.