Remove POST_comments
Now that we're not saving sort preferences when you select a new sort from the
menu in comment threads, we don't really need to have a `POST` endpoint for
comments; all that did was handle the save and redirect.

Untangling how this menu got created took me some time, so try to hold on while
I explain what's happening here.

`CommentSortMenu`, which handles, as you may guess, the sort menu for comments,
had `button_cls` set to `PostButton`.  That in turn defines its `_style` as
"post".  And *that* has the effect of `post()`, defined in
`templates/navbutton.html`, being the function used to render the button (which
we do in `templates/navmenu.html`'s `dropdown()`).  And `post()` calls, not
surprisingly, `post_link()`, which builds up a `POST` form.

Ok, we'll pause for a minute to let that call chain seep in a little bit, at
least.  Good?

So now if we no longer set `button_cls`, we default to the parent's
`button_cls`, which is `QueryButton`.  `QueryButton`, as its name implies, is
designed to pass arguments in the query string, and does so through the normal
`plain_link()` (called from the template's `plain()`, which is used because the
`_style` is set to "plain").

Aside: we don't want to get rid of `PostButton` entirely because it's still
used by `ControversyTimeMenu`.