CommentPane: just get the comments when checking if we can use cache.
The cached version of a CommentPane is valid only if the user is not the
author of any of the comments. Previously we would retrieve and build
the comment tree for the user to check if we could use the cache, and then
on a cache miss we'd re-retrieve and re-build the tree for a logged out user
to create a version suitable for caching.

This commit changes it so the CommentBuilder can retrieve the comments that
will be shown so we can check if the user authored any of them. Then if we
can use the cache and there's a cache miss we can build the rendered tree
for the logged out user without re-retrieving, and the building happens only
at this stage.