Image previews: prefer direct image links
As requested by @dforsyth, when extracting links from self-posts for image
previews, we'll now use an imgur or image-looking link if available, overriding
the default behavior of choosing the first link in the post text.  This helps
us deal with AMAs in particular, where sometimes posters will link to their
website, YouTube videos, etc. and have a "proof image" at the bottom of the
post.