reddit header: format karma number
In the top-right of every logged-in reddit page, we display the user's link
karma.  This number has been unformatted, which makes it a bit difficult to
read at a glance as you gain a fair amount of karma.  Now we'll format it with
periods or commas, according to the locale.