Add loggedin cache poisoning detection using a canary cookie
There've been a few instances in the past where we poisoned downstream
caches with loggedin responses for several hours before realizing it was
happening. Each time we only realized it once a user reported that they
were being logged in as someone else.

This PR attempts to make it easier to detect upticks in cache poisoning
by adding a cache poisoning canary[0] cookie. The canary cookie is a
random, non-identifying, persistent cookie that is included in every
request and sent back with every loggedin HTML response. If the canary
in the response is different from the one we would have sent, we know
that cache poisoning occurred.

If the client detects it has been served an improperly cached page,
it tries to gather extra data about the page (such as some of the
headers sent with response for the current page.) With those, we can
often figure out _how_ the cache was poisoned (did CF cache it, did
an intermediary proxy mess with the `Cache-Control` headers?) We can
also sometimes find out what was served with the poisoned response,
and if the victim is now likely able to perform actions as the poisoner.)

We also keep track of what cache policy was used, so we can experiment
with incremental rollout of new cache policies and see how they affect
the incidence rate of poisoned caches.

We keep track of the overall cache poisoning report rate in tallier,
and send a more detailed report to event-collector so we can figure
out what's causing all the poisonings and who was affected after the fact.

[0]: https://en.wikipedia.org/wiki/Buffer_overflow_protection#Canaries

Conflicts:
	r2/r2/lib/template_helpers.py