Image previews: add url signing capabilities
We've been trying to avoid using signed urls for preview images simply because
it's easier to not.

However, during the business discussions with imgix, they raised a concern
about this; namely, the pricing structure we've arranged would add nothing (or
little, I'm not exactly sure) to our bill if our users take advantage of
imgix's url parameters to create new images based off an existing source.

So, we'll now turn on signing for production.