Image previews: correctly sign resized images
When creating the string to hash as part of imgix's url-signing process, we
were previously erroneously ignoring any updates made to the query params.
This caused any resized images to fail the signature check.

Now we look at the `UrlParser`'s `query_dict` attribute, which is updated by
our previous calls to `update_query()`, rather than plain ol' `query`, which is
merely a record of the original query parameters found in the url passed into
the constructor.  I'll change that API to make more sense, but do so separately
from this.