Fix header override when stylesheet doesn't exist
With the reddit themes release, the default sr style and header overrides
the subreddit style if there is no stylesheet (even if there is a header).
This change considers a header as being part of the style and won't override
the current sr style for the reason of there being no stylesheet. This header
image can be disabled with the "show this subreddit's theme" toggle.