Q&A sort: fix caching bug
In certain cases, we'll try to load in comments for which we can't find a
cached sort value.  When that happens, we calculate the value on-the-fly and
then store it.

When this happens, we've been pulling the comment without any of its extra
attributes.  This is fine for most sorts, but for the Q&A sort, we look at more
than just the ups and downs (notably, `author_id`).  Since we weren't getting
these attributes, we'd run into an error when trying to use them.

Now we fetch the extra data for comments when using the Q&A sort.