Footer: switch mobile sites
We've launched the new new mobile web experience - yay!  Since we switched over
the mobile hint for Google in reddit/reddit@63f054d, I figured it was probably
time to do so in the footer as well.

Oh, and it redirects all http traffic to https, so we should just point to that
directly.