Add support for 'ref' tags and add to trending subreddits.
A "ref tag" is a query parameter we may pass to ourselves to let
us know where a user came from, and more specifically, what part
of the page they clicked on to get there. This allows us to know,
for example, if users are clicking through from trending
subreddits or if they're just clicking on the subreddit from a
link. This helps us know if features are used/valuable or not.

A ref tag will be passed through and then removed from the URL via
javascript post-load to make a clean URL that's nice for
copy/pasting elsewhere, and doesn't give false positives if they
do paste it somewhere else.