AutoMod: fix method of setting approve_banned
Previously, approve_banned was set to True only if an exact check for
"name" was being done inside the author target. This excluded other
possibilities that also checked against name, such as "~name",
"name (regex)", and so on. This is solved by looking in match_fields
instead to determine whether name is being checked against in any way.