Add link to AutoMod config in moderation tools
This link will only appear once the subreddit has already created an
AutoMod config wiki page, to keep it as a bit more of an "advanced
feature" that people have to go out of their way to discover. If we're
more comfortable with the level of AutoModerator information/education
in the future, we can change this to always show in the list.