Image previews: add nsfw variant
We'd like to be able to show some sort of preview for NSFW images, rather than
just a placeholder image.  Thanks to imgix, it's easy for us to create
pixellated versions of the preview images for app developers to display.

We went back and forth a bit on whether a NSFW image should have the censored
version in the main image block (with the uncensored version as a variant), or
have the censored version be the variant.  We settled on the latter because an
app that doesn't build in NSFW link support will perhaps just shock a few
users, as opposed to preventing NSFW-seeking users from obtaining the content
they want.  It's also less of a backwards-compatibility change.