AutoMod: catch and log invalid config errors
In the case of changes that cause previously-valid rule definitions to
start raising an exception, there may be some configurations that were
saved before the change was made, and so now contain an invalid
definition. This will cause the automod queue consumer(s) to hang if an
item from any of those subreddits needs to be checked, since it will be
unable to initialize the Ruleset to be able to process it. This commit
makes it so that items from subreddits with invalid configurations will
be skipped and an error logged.