AutoMod: force domain and add context to permalink
The {{permalink}} placeholder (which is auto-included in messages and
modmails) has been using relative links, but some apps don't handle
these properly, so it's better to go back to using a full link with the
domain.

In addition, this makes it so that non-top-level comments will have
context=3 added on their permalinks.