Track subreddit participation directly
The previous method of determining whether a user had interacted with a
particular subreddit (to decide whether to send them a "you've been
banned" message or not) was by checking to see whether they had any
karma (positive or negative) in the subreddit, or if they were currently
subscribed to it. This missed a fair number of cases, most significantly
that making self-posts would not count as having interacted with the
subreddit, since they have no effect on karma.

This commit uses a DenormalizedRelation to track whether a user has
interacted with the subreddit (actually *when* they last interacted with
it, though that data is currently unused). It adds one new type of
action that is considered interaction - sending a modmail to the
subreddit.  It also removes the error/warning from the ban page about
trying to send a ban message to someone that has never interacted with
the subreddit, since that should no longer really be a concern.